Extract `I18nContext`
from `provider` into `context`
