+++ src/app/i18n/mod.rs
@@ -1,3 +1,4 @@
+mod context;
 mod de;
 mod en;
 mod hi;
@@ -6,6 +7,7 @@ mod zh_hant;
 
 mod types;
 
+pub use context::*;
 pub use types::*;
 
 #[allow(dead_code)]
